GLD: fix and move into one command
